fixed id to int from long

+++ src/main/java/com/codeup/adlister/models/Households.java
@@ -1,7 +1,7 @@
 package com.codeup.adlister.models;
 
 public class Households {
-    private long id;
+    private int id;
     private String username;
     private String email;
     private String password;
@@ -15,13 +15,13 @@ public class Households {
         this.password = password;
     }
 
-    public Households(long id, String username, String email, String password) {
+    public Households(int id, String username, String email, String password) {
         this.id = id;
         this.username = username;
         this.email = email;
         this.password = password;
     }
-    public Households(long id, String username, String email, String password, String name) {
+    public Households(int id, String username, String email, String password, String name) {
         this.id = id;
         this.username = username;
         this.email = email;
@@ -41,7 +41,7 @@ public class Households {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
